XMP-40: move secrets initialization into utils;
